+++ lib/utils/amiibo_category.dart
@@ -1,0 +1,13 @@
+enum AmiiboCategory{
+  All,
+  Owned,
+  Wishlist,
+  Figures,
+  FigureSeries,
+  Cards,
+  CardSeries,
+  Name,
+  AmiiboSeries,
+  Game,
+  Custom
+}